+++ src/Farmer/Builders/Builders.Functions.fs
@@ -284,6 +284,7 @@ type FunctionsConfig =
                   WorkerCount = 0
                   MaximumElasticWorkerCount = None
                   OperatingSystem = this.CommonWebConfig.OperatingSystem
+                  ZoneRedundant = None
                   Tags = this.Tags }
             | _ ->
                 ()