+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 ## vNext
 * WebApps/Functions: Don't turn on Logging Extension for Linux App Service.
 * WebApps/Functions: Fix .NET 5/6 on Linux deployments.
+* ServicePlan/WebApp: Support for enabling ZoneDedundant
 
 ## 1.6.25
 * CosmosDb: Add support for serverless capacity mode.